core: fixes improper handling of file under Python 2 (#32)
File reading now reflect PEP 263 which ensure that the lexer is always passed utf-8 encoded string. Remove the need to pass the encoding to the lexer. Both import_hooks and runner logic have tested to work under Python 2 and 3 and to produce the correct result (test string is 'I can contain unicode characters: αΨγ 😀⏰🏔 𠁒𠁝𠁖')